requirements.txt
# Manipulação de arquivos e EXIF
exiftool-wrapper
pyexiftool

# Tipagem e utilitários
typing-extensions
tqdm

# Processamento de imagem e vídeo
opencv-python
moviepy
numpy<2.3.0
Pillow
pillow-heif
pyheif

# Texto e narração
transformers
gTTS

# Geolocalização e mapas
geopy
staticmap
folium
selenium
chromedriver-autoinstaller

# CLI e TUI (opcional)
rich
